ai_service/requirements.txt
# Web Framework & Server
fastapi>=0.136.0
uvicorn[standard]>=0.35.0

# AI & LLM (Switched from xAI to Groq)
langchain-groq>=1.1.2
google-genai>=1.0.0
langchain-core>=0.3.0

# Vector Database
qdrant-client>=1.17.1

# Configuration & Validation
python-dotenv>=1.2.2
pydantic>=2.10.0
pydantic-settings>=2.14.0